requirements.txt
Django==2.2.1
django-bootstrap-form==3.4
djangorestframework==3.9.4
html5lib==1.0.1
Pillow==6.0.0
pkg-resources==0.0.0
PyPDF2==1.26.0
pytz==2019.1
reportlab==3.5.21
six==1.12.0
sqlparse==0.3.0
webencodings==0.5.1
xhtml2pdf==0.2.3
xlwt==1.3.0


